+++ .gitignore
@@ -264,6 +264,8 @@ redis-demo/springboot-redis-distribute-lock/src/main/resources/application-local
 
 /system-design-demo/MyLimiter/.idea
 /system-design-demo/MyLimiter/target
+
+
 /system-design-demo/TokenCheck/.idea
 /system-design-demo/TokenCheck/target
 /system-design-demo/TokenCheck/src/main/resources/application-dev.yml
