+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day31/Solution377_02.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan01.day31;
+package com.noob.algorithm.daily.plan01.archive.day31;
 
 import java.util.Arrays;
