+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day33/Solution337_03.java
@@ -1,9 +1,6 @@
-package com.noob.algorithm.daily.plan01.day33;
+package com.noob.algorithm.daily.plan01.archive.day33;
 
 import com.noob.algorithm.daily.base.TreeNode;
-
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * 🟡 337 打家劫舍III - https://leetcode.cn/problems/house-robber-iii/description/
