+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan01/archive/day34/Solution123_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan01.day34;
+package com.noob.algorithm.daily.plan01.archive.day34;
 
 /**
  * 🔴 123 买卖股票的最佳时机III - https://leetcode.cn/problems/best-time-to-buy-and-sell-stock-iii/description/
